separate out flags implementation from FilesystemDefaultImplementationTrait
